requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=dev.txt dev.in
#
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==24.2.0
    # via aiohttp
blinker==1.8.2
    # via flask
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.3.2
    # via
    #   pdfminer-six
    #   requests
click==8.1.7
    # via flask
cryptography==43.0.1
    # via pdfminer-six
dataclasses-json==0.6.7
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
faiss-cpu==1.7.4
    # via -r base.in
flask==3.0.3
    # via -r base.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.1.0
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
jiter==0.5.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.1.20
    # via -r base.in
langchain-community==0.0.38
    # via
    #   -r base.in
    #   langchain
langchain-core==0.1.52
    # via
    #   -r base.in
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.0.8
    # via -r base.in
langchain-text-splitters==0.0.2
    # via langchain
langsmith==0.1.123
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.22.0
    # via dataclasses-json
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
openai==1.44.1
    # via
    #   -r base.in
    #   langchain-openai
orjson==3.10.7
    # via langsmith
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pdfminer-six==20231228
    # via pdfplumber
pdfplumber==0.11.4
    # via -r base.in
pillow==10.4.0
    # via pdfplumber
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.23.4
    # via pydantic
pypdfium2==4.30.0
    # via pdfplumber
python-dotenv==1.0.1
    # via -r base.in
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   tiktoken
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.35
    # via
    #   langchain
    #   langchain-community
tenacity==8.5.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.7.0
    # via langchain-openai
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.3
    # via requests
werkzeug==3.0.4
    # via flask
yarl==1.11.1
    # via aiohttp
